# Problem: block-stm has bad worst case performance