Solution:
- support optional pre-estimate logic to improve worst case performance, in the cost of optimal case performance.

## Benchmark Results

* account 1 (worst case)
```
$ benchstat /tmp/before /tmp/after
goos: darwin
goarch: arm64
pkg: github.com/crypto-org-chain/cronos/v2/app
                                │  /tmp/before  │             /tmp/after             │
                                │    sec/op     │   sec/op     vs base               │
ERC20Transfer/memiavl-stm-16-16   1533.1m ± 11%   677.5m ± 2%  -55.81% (p=0.002 n=6)

                                │  /tmp/before  │             /tmp/after              │
                                │     B/op      │     B/op      vs base               │
ERC20Transfer/memiavl-stm-16-16   7316.8Mi ± 5%   532.6Mi ± 0%  -92.72% (p=0.002 n=6)

                                │  /tmp/before  │             /tmp/after             │
                                │   allocs/op   │  allocs/op   vs base               │
ERC20Transfer/memiavl-stm-16-16   120.871M ± 5%   6.488M ± 0%  -94.63% (p=0.002 n=6)
```

* account 100
```
goos: darwin
goarch: arm64
pkg: github.com/crypto-org-chain/cronos/v2/app
                                │ /tmp/before │             /tmp/after             │
                                │   sec/op    │   sec/op     vs base               │
ERC20Transfer/memiavl-stm-16-16   122.5m ± 1%   141.3m ± 3%  +15.29% (p=0.002 n=6)

                                │ /tmp/before  │          /tmp/after           │
                                │     B/op     │     B/op      vs base         │
ERC20Transfer/memiavl-stm-16-16   493.8Mi ± 0%   493.1Mi ± 0%  ~ (p=0.485 n=6)

                                │ /tmp/before │            /tmp/after             │
                                │  allocs/op  │  allocs/op   vs base              │
ERC20Transfer/memiavl-stm-16-16   6.259M ± 0%   6.195M ± 0%  -1.02% (p=0.002 n=6)
```

👮🏻👮🏻👮🏻 !!!! REFERENCE THE PROBLEM YOUR ARE SOLVING IN THE PR TITLE AND DESCRIBE YOUR SOLUTION HERE !!!! DO NOT FORGET !!!! 👮🏻👮🏻👮🏻


# PR Checklist:

- [ ] Have you read the [CONTRIBUTING.md](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md)?
- [ ] Does your PR follow the [C4 patch requirements](https://rfc.zeromq.org/spec:42/C4/#23-patch-requirements)?
- [ ] Have you rebased your work on top of the latest master? 
- [ ] Have you checked your code compiles? (`make`)
- [ ] Have you included tests for any non-trivial functionality?
- [ ] Have you checked your code passes the unit tests? (`make test`)
- [ ] Have you checked your code formatting is correct? (`go fmt`)
- [ ] Have you checked your basic code style is fine? (`golangci-lint run`)
- [ ] If you added any dependencies, have you checked they do not contain any known vulnerabilities? (`go list -json -m all | nancy sleuth`)
- [ ] If your changes affect the client infrastructure, have you run the integration test?
- [ ] If your changes affect public APIs, does your PR follow the [C4 evolution of public contracts](https://rfc.zeromq.org/spec:42/C4/#26-evolution-of-public-contracts)?
- [ ] If your code changes public APIs, have you incremented the crate version numbers and documented your changes in the [CHANGELOG.md](https://github.com/crypto-org-chain/chain-main/blob/master/CHANGELOG.md)?
- [ ] If you are contributing for the first time, please read the agreement in [CONTRIBUTING.md](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md) now and add a comment to this pull request stating that your PR is in accordance with the [Developer's Certificate of Origin](https://github.com/crypto-org-chain/chain-main/blob/master/CONTRIBUTING.md#developer-certificate-of-originn).

Thank you for your code, it's appreciated! :)



<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
	- Introduced a pre-estimate block-stm option to enhance worst-case performance.
	- Added a new parameter for block-stm executor configuration, improving application behavior during execution.

- **Bug Fixes**
	- No bugs introduced or existing functionalities altered.

- **Documentation**
	- Updated CHANGELOG.md to reflect recent enhancements and changes.

- **Chores**
	- Updated dependency versions to ensure alignment with the latest libraries and improvements.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->